test:전략 패턴을 위해 뺄셈 테스트를 StringCalculatorTest에서 SubtractOperatorTest로 리팩토링
